test(Catalog): Change namespaces for existing tests
Having 'Tests' ending in each part of namespace looks a bit clumsy.
So I decided to remove it.
In order to avoid mistake with the ambiguity between unit test
namespace and name of a class that is being tested, I added a ne
using condition, that calls the class that is being tested as a target.
I suggest using this idea in other tests in order to have all tests
following the one convention.