fix(Catalog): Rearrange Comments DataAccessor
Made changes to the Comments DataAcessor
to sync changes in the branch with the dev.
As long as Comment already contain ProductId as its property, we don't need to
pass productId itself as an additional
parameter for the method most of the time.

However, when GetAll method for
Comments is not needed as long as they
are not making sense without a Product.
For that reason I created new child of ISqlDataAccessor for that purpose

I've also adjusted existing tests for the code consistency.