test(intg): Add integration tests for GetProductAsync
In order to test how integration tests are working with
exceptions, I've created test that will cause NotFound
Status code. For API to work as expected, I've made changes to the GetProductAsync method of the
ProductsController.

I also changed azure-pipelines.yml, so that it does not
invoke integration tests, that will inevitably fail without
proper database setup and should not be run
automatically