feat(wbs): Redesign gateway using Ocelot
Previous version of API Gateway did its job in a wrong
way, thus making it hard to work with and giving
nothing to our advantage.
Now it was created to work using Ocelot - lightweight
API Gateway middleware.

Previously created logic was removed from the project.
All tests for this logic was also deleted.