refactor(catalog): Refactor AddComments method and usages
AddComment method that accepts Comment as arg may
not be the best solution - it allows to pass Comment with
the wrong productId. AddComment(authorId, message)
seems to be less error-prone and more natural option
Other changes in this commit made to fit new method
signature.

AddCommentsFromDtoList and AddSpecificationsFromDtoList
methods are taken out according to the DRY principle